+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add/retrieve generated data from utilities #225
 - Add `skip_tests` option to `dockci.yaml` for use with utilites #223
 - Specify `Dockerfile` to use in `dockci.yaml` #227
+- Allow override of TLS params per Docker host #235
 
 ### v0.0.6
 - Significant decrease in browser load on jobs with error stage #215